requirements.txt
quart==0.16.1
aiomysql
aioneo4j
huggingface-hub
aiofiles
aiorwlock
apscheduler
Quart-CORS==0.5.0
jinja2==3.0.3
urllib3
idna
certifi
wheel
numpy
pandas
matplotlib
flask
scikit-learn
treelib>=1.6.1
torch
torchvision
opencv-python 
jupyter
notebook
tokenizers==0.10.2
gensim
graphviz  
transformers==4.9.2
pytorch_transformers
jieba
pyhanlp
py2neo==2021.0.1
pymysql
DBUtils
werkzeug==2.0.2
openpyxl
fasttext
pkuseg
networkx
pinyin
aiomysql
aioneo4j
aiofiles
aiorwlock
apscheduler
pyahocorasick
checklist
allennlp==2.7.0
Levenshtein
fuzzywuzzy